+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3] - 2025-10-11
 - core: pull git repos if already cloned
 - core: permit to change repos folder path
 - export: add suport for remote export of reports with rclone